+++ SourceryTests/Models/TypedSpec.generated.swift
@@ -26,6 +26,11 @@ class TypedSpec: QuickSpec {
                 expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
             }
 
+            it("can report array type via KVC") {
+                let sut = AssociatedValue(typeName: TypeName("[Int]"))
+                expect(sut.value(forKeyPath: "isArray") as? Bool).to(equal(true))
+            }
+
             it("can report actual type name via KVC") {
                 let sut = AssociatedValue(typeName: TypeName("Alias"))
                 expect(sut.value(forKeyPath: "actualTypeName") as? TypeName).to(equal(TypeName("Alias")))
@@ -51,6 +56,11 @@ class TypedSpec: QuickSpec {
             it("can report closure type via KVC") {
                 let sut = MethodParameter(typeName: TypeName("(Int) -> (Int)"))
                 expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
+            }
+
+            it("can report array type via KVC") {
+                let sut = AssociatedValue(typeName: TypeName("[Int]"))
+                expect(sut.value(forKeyPath: "isArray") as? Bool).to(equal(true))
             }
 
             it("can report actual type name via KVC") {
@@ -80,6 +90,11 @@ class TypedSpec: QuickSpec {
                 expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
             }
 
+            it("can report array type via KVC") {
+                let sut = AssociatedValue(typeName: TypeName("[Int]"))
+                expect(sut.value(forKeyPath: "isArray") as? Bool).to(equal(true))
+            }
+
             it("can report actual type name via KVC") {
                 let sut = AssociatedValue(typeName: TypeName("Alias"))
                 expect(sut.value(forKeyPath: "actualTypeName") as? TypeName).to(equal(TypeName("Alias")))
@@ -107,6 +122,11 @@ class TypedSpec: QuickSpec {
                 expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
             }
 
+            it("can report array type via KVC") {
+                let sut = AssociatedValue(typeName: TypeName("[Int]"))
+                expect(sut.value(forKeyPath: "isArray") as? Bool).to(equal(true))
+            }
+
             it("can report actual type name via KVC") {
                 let sut = AssociatedValue(typeName: TypeName("Alias"))
                 expect(sut.value(forKeyPath: "actualTypeName") as? TypeName).to(equal(TypeName("Alias")))
@@ -132,6 +152,11 @@ class TypedSpec: QuickSpec {
             it("can report closure type via KVC") {
                 let sut = Variable(typeName: TypeName("(Int) -> (Int)"))
                 expect(sut.value(forKeyPath: "isClosure") as? Bool).to(equal(true))
+            }
+
+            it("can report array type via KVC") {
+                let sut = AssociatedValue(typeName: TypeName("[Int]"))
+                expect(sut.value(forKeyPath: "isArray") as? Bool).to(equal(true))
             }
 
             it("can report actual type name via KVC") {